feat(utils): Move codecs and types to proper packages, to prevent circular dependencies later
