+++ apps/dashboard-for-dapps/package.json
@@ -6,8 +6,8 @@
     "@chakra-ui/react": "^3.15.1",
     "@idos-network/client": "workspace:*",
     "@idos-network/core": "workspace:*",
-    "@stablelib/base64": "^2.0.1",
-    "@stablelib/utf8": "^2.0.1",
+    "@idos-network/utils": "workspace:*",
+    "@idos-network/credentials": "workspace:*",
     "@tanstack/query-sync-storage-persister": "catalog:tanstack",
     "@tanstack/react-query": "catalog:tanstack",
     "@tanstack/react-query-persist-client": "catalog:tanstack",
