+++ apps/dashboard-for-dapps/src/components/secret-key-prompt.tsx
@@ -1,5 +1,5 @@
 import { Stack, Text } from "@chakra-ui/react";
-import type { idOSCredential } from "@idos-network/core";
+import type { idOSCredential } from "@idos-network/credentials";
 import { useRef, useState } from "react";
 
 import {
