+++ apps/dashboard-for-dapps/src/utils.ts
@@ -1,4 +1,4 @@
-import { base64Decode, utf8Decode } from "@idos-network/core/codecs";
+import { base64Decode, utf8Decode } from "@idos-network/utils/codecs";
 import nacl from "tweetnacl";
 
 export function decrypt(b64FullMessage: string, b64SenderPublicKey: string, secretKey: string) {
