+++ apps/embedded-wallet/tsconfig.app.json
@@ -26,5 +26,5 @@
     "noFallthroughCasesInSwitch": true,
     "noUncheckedSideEffectImports": true
   },
-  "include": ["src"]
+  "include": ["src", "../../packages/@credentials/src/types.d.ts"]
 }
