+++ apps/idos-data-dashboard/package.json
@@ -8,6 +8,8 @@
     "@emotion/styled": "^11.13.0",
     "@gemwallet/api": "^3.8.0",
     "@idos-network/client": "workspace:*",
+    "@idos-network/credentials": "workspace:*",
+    "@idos-network/utils": "workspace:*",
     "@idos-network/core": "workspace:*",
     "@near-wallet-selector/core": "catalog:near",
     "@near-wallet-selector/here-wallet": "catalog:near",
@@ -15,6 +17,8 @@
     "@near-wallet-selector/modal-ui": "catalog:near",
     "@stellar/stellar-base": "catalog:stellar",
     "@tanstack/react-query": "catalog:tanstack",
+    "@reown/appkit": "catalog:reown",
+    "@reown/appkit-adapter-wagmi": "catalog:reown",
     "@tanstack/react-query-devtools": "catalog:tanstack",
     "@web3icons/react": "^4.0.14",
     "@web3modal/wagmi": "^4.2.3",
