+++ apps/idos-data-dashboard/src/routes/dashboard/credentials/types.ts
@@ -1,4 +1,4 @@
-import type { idOSCredential } from "@idos-network/core";
+import type { idOSCredential } from "@idos-network/credentials";
 
 export type idOSCredentialWithShares = idOSCredential & {
   shares: string[];
