+++ apps/idos-data-dashboard/tsconfig.json
@@ -25,6 +25,6 @@
       "@/*": ["src/*"]
     }
   },
-  "include": ["src"],
+  "include": ["src", "../../packages/@credentials/src/types.d.ts"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
