+++ apps/idos-enclave/package.json
@@ -3,8 +3,9 @@
   "description": "idOS Enclave",
   "version": "0.1.0",
   "dependencies": {
-    "@idos-network/utils": "workspace:*",
+    "@idos-network/credentials": "workspace:*",
     "@idos-network/core": "workspace:*",
+    "@idos-network/utils": "workspace:*",
     "@partisiablockchain/abi-client": "^5.207.0",
     "@partisiablockchain/blockchain-api-transaction-client": "^5.149.0",
     "@preact/signals": "^2.0.2",
