+++ apps/isle/package.json
@@ -8,6 +8,7 @@
     "@emotion/react": "^11.13.0",
     "@fontsource-variable/inter": "^5.2.5",
     "@idos-network/core": "workspace:*",
+    "@idos-network/credentials": "workspace:*",
     "@sanity/comlink": "^3.0.1",
     "@uidotdev/usehooks": "^2.4.1",
     "motion": "^12.4.1",
