+++ apps/isle/tsconfig.app.json
@@ -24,5 +24,5 @@
     "noFallthroughCasesInSwitch": true,
     "noUncheckedSideEffectImports": true
   },
-  "include": ["src"]
+  "include": ["src", "../../packages/@credentials/src/types.d.ts"]
 }
