+++ examples/consumer-and-issuer/package.json
@@ -8,6 +8,8 @@
     "@idos-network/consumer": "workspace:*",
     "@idos-network/controllers": "workspace:*",
     "@idos-network/core": "workspace:*",
+    "@idos-network/credentials": "workspace:*",
+    "@idos-network/utils": "workspace:*",
     "@idos-network/issuer": "workspace:*",
     "@nanostores/react": "^0.8.4",
     "@near-wallet-selector/core": "catalog:near",
