+++ examples/pay-demo/package.json
@@ -2,26 +2,21 @@
   "name": "pay-demo",
   "version": "0.0.1",
   "dependencies": {
-    "@digitalbazaar/ed25519-signature-2020": "^5.4.0",
-    "@digitalbazaar/ed25519-verification-key-2020": "^4.2.0",
-    "@digitalbazaar/vc": "^7.2.0",
     "@idos-network/client": "workspace:*",
     "@idos-network/consumer": "workspace:*",
+    "@idos-network/utils": "workspace:*",
     "@react-router/node": "^7.6.2",
     "@react-router/serve": "^7.6.2",
     "@reown/appkit": "^1.7.8",
     "@reown/appkit-adapter-wagmi": "^1.7.8",
-    "@statelyai/inspect": "^0.4.0",
     "@transak/transak-sdk": "^3.2.0",
     "@vercel/react-router": "^1.1.2",
     "@xstate/react": "^5.0.5",
     "ascii85": "^1.0.2",
-    "base85": "^3.1.0",
     "countries-list": "^3.1.1",
     "ethers": "catalog:ethereum",
     "file-type": "^21.0.0",
     "isbot": "^5.1.28",
-    "jsonld-document-loader": "^2.3.0",
     "jsonwebtoken": "^9.0.2",
     "react": "catalog:react",
     "react-dom": "catalog:react",
