+++ packages/@core/package.json
@@ -6,18 +6,10 @@
     "@gemwallet/api": "^3.8.0",
     "@kwilteam/kwil-js": "catalog:core",
     "@near-wallet-selector/core": "catalog:near",
-    "@stablelib/base64": "catalog:core",
-    "@stablelib/binary": "catalog:core",
-    "@stablelib/bytes": "catalog:core",
-    "@stablelib/hex": "catalog:core",
-    "@stablelib/sha256": "catalog:core",
-    "@stablelib/utf8": "catalog:core",
     "@stellar/stellar-sdk": "catalog:stellar",
     "@idos-network/utils": "workspace:*",
+    "@idos-network/credentials": "workspace:*",
     "@wagmi/core": "2.17.3",
-    "borsh": "catalog:core",
-    "bs58": "catalog:core",
-    "es-toolkit": "catalog:core",
     "ethers": "catalog:ethereum",
     "near-api-js": "catalog:near",
     "ripple-binary-codec": "catalog:xrp",
@@ -31,7 +23,6 @@
   },
   "exports": {
     ".": "./src/index.ts",
-    "./codecs": "./src/codecs/index.ts",
     "./signature-verification": "./src/signature-verification/index.ts",
     "./cryptography": "./src/cryptography/index.ts",
     "./types": "./src/types/index.ts",
