+++ packages/@core/src/signature-verification/near.ts
@@ -1,6 +1,6 @@
+import { bs58Decode, hexDecode, sha256Hash } from "@idos-network/utils/codecs";
 import { deserialize as borshDeserialize, serialize as borshSerialize } from "borsh";
 import nacl from "tweetnacl";
-import { bs58Decode, hexDecode, sha256Hash } from "../codecs";
 
 interface NEP413Payload {
   tag: number;
