+++ packages/@core/src/utils/index.ts
@@ -1,7 +1,7 @@
+import type { InsertableIDOSCredential } from "@idos-network/credentials";
+import { base64Decode, base64Encode, hexEncode, utf8Encode } from "@idos-network/utils/codecs";
 import invariant from "tiny-invariant";
 import nacl from "tweetnacl";
-import { base64Decode, base64Encode, hexEncode, utf8Encode } from "../codecs";
-import type { InsertableIDOSCredential } from "../types";
 
 export async function buildInsertableIDOSCredential(
   userId: string,
