+++ packages/client/src/enclave/metamask-snap-enclave.ts
@@ -1,4 +1,4 @@
-import type { idOSCredential } from "@idos-network/core";
+import type { idOSCredential } from "@idos-network/credentials";
 import type {
   DiscoverUserEncryptionPublicKeyResponse,
   EnclaveOptions,
