+++ packages/consumer/package.json
@@ -13,7 +13,8 @@
     "@digitalbazaar/ed25519-verification-key-2020": "catalog:core",
     "@digitalbazaar/vc": "catalog:core",
     "@kwilteam/kwil-js": "catalog:core",
-    "base85": "catalog:core",
+    "@idos-network/utils": "workspace:*",
+    "base85": "catalog:utils",
     "jsonld-document-loader": "catalog:core",
     "ripple-keypairs": "catalog:xrp",
     "tweetnacl": "catalog:core",
