+++ packages/issuer/package.json
@@ -12,9 +12,10 @@
     "@digitalbazaar/ed25519-signature-2020": "catalog:core",
     "@digitalbazaar/ed25519-verification-key-2020": "catalog:core",
     "@digitalbazaar/vc": "catalog:core",
+    "@idos-network/utils": "workspace:*",
     "@kwilteam/kwil-js": "catalog:core",
-    "base85": "catalog:core",
-    "es-toolkit": "catalog:core",
+    "base85": "catalog:utils",
+    "es-toolkit": "catalog:utils",
     "jsonld-document-loader": "catalog:core",
     "ripple-keypairs": "catalog:xrp",
     "tweetnacl": "catalog:core",
