+++ packages/issuer/src/services/credential.service.ts
@@ -1,13 +1,4 @@
 import {
-  base64Decode,
-  base64Encode,
-  encryptContent,
-  hexEncode,
-  hexEncodeSha256Hash,
-  type KwilActionClient,
-  utf8Encode,
-} from "@idos-network/core";
-import {
   type CreateCredentialByDelegatedWriteGrantParams,
   createCredentialByDelegatedWriteGrant,
   type EditCredentialAsIssuerParams,
@@ -15,7 +6,16 @@ import {
   getCredentialIdByContentHash,
   getSharedCredential,
 } from "@idos-network/core/kwil-actions";
-import type { idOSCredential } from "@idos-network/core/types";
+import type { KwilActionClient } from "@idos-network/core/kwil-infra";
+import type { idOSCredential } from "@idos-network/credentials";
+import {
+  base64Decode,
+  base64Encode,
+  encryptContent,
+  hexEncode,
+  hexEncodeSha256Hash,
+  utf8Encode,
+} from "@idos-network/utils/codecs";
 import invariant from "tiny-invariant";
 import nacl from "tweetnacl";
