+++ packages/utils/tsdown.config.ts
@@ -1,6 +1,6 @@
 import { defineConfig } from "tsdown";
 
 export default defineConfig({
-  entry: ["./src/encryption/index.ts", "./src/store/index.ts"],
+  entry: ["./src/encryption/index.ts", "./src/store/index.ts", "./src/codecs/index.ts"],
   dts: true,
 });
