+++ pnpm-workspace.yaml
@@ -10,6 +10,13 @@ catalogs:
     "@digitalbazaar/vc": "^7.2.0"
     "@gemwallet/api": "^3.8.0"
     "@kwilteam/kwil-js": "0.9.4"
+    "jsonld-document-loader": "^2.3.0"
+    "tweetnacl": "^1.0.3"
+    "tiny-invariant": "^1.3.3"
+    "web-streams-polyfill": "^4.1.0"
+    "zod": "^3.25.64"
+
+  utils:
     "@stablelib/base64": "^2.0.1"
     "@stablelib/binary": "^2.0.1"
     "@stablelib/bytes": "^2.0.1"
@@ -19,12 +26,7 @@ catalogs:
     "base85": "^3.1.0"
     "borsh": "^2.0.0"
     "bs58": "^6.0.0"
-    "jsonld-document-loader": "^2.3.0"
     "es-toolkit": "^1.39.0"
-    "tweetnacl": "^1.0.3"
-    "tiny-invariant": "^1.3.3"
-    "web-streams-polyfill": "^4.1.0"
-    "zod": "^3.25.64"
 
   ethereum:
     "wagmi": "^2.14.16"
